+++ invokeai/backend/__init__.py
@@ -1,6 +1,7 @@
 '''
 Initialization file for invokeai.backend
 '''
-from .invoke_ai_web_server import InvokeAIWebServer
+# this is causing circular import issues
+# from .invoke_ai_web_server import InvokeAIWebServer
