+++ invokeai/backend/generator/base.py
