+++ invokeai/backend/generator/embiggen.py
