+++ invokeai/backend/generator/omnibus.py
