+++ invokeai/backend/models/autoencoder.py
