+++ invokeai/backend/models/diffusion/__init__.py
@@ -1,0 +1,6 @@
+'''
+Initialization file for invokeai.models.diffusion
+'''
+from .shared_invokeai_diffusion import InvokeAIDiffuserComponent, PostprocessingSettings
+from .cross_attention_control import InvokeAICrossAttentionMixin
+from .cross_attention_map_saving import AttentionMapSaver
