+++ invokeai/backend/models/diffusion/cross_attention_control.py
