+++ invokeai/backend/models/diffusion/ddim.py
