+++ invokeai/backend/models/diffusion/ddpm.py
