+++ invokeai/backend/models/diffusion/shared_invokeai_diffusion.py
