--- invokeai/models/diffusion/__init__.py
@@ -1,4 +1,0 @@
-'''
-Initialization file for invokeai.models.diffusion
-'''
-from .shared_invokeai_diffusion import InvokeAIDiffuserComponent
