+++ ldm/invoke/merge_diffusers.py
@@ -23,10 +23,9 @@ from omegaconf import OmegaConf
 from ldm.invoke.config.widgets import FloatTitleSlider
 from ldm.invoke.globals import (Globals, global_cache_dir, global_config_file,
                                 global_models_dir, global_set_root)
-from invokeai.models import ModelManager
+from invokeai.backend.models import ModelManager
 
 DEST_MERGED_MODEL_DIR = "merged_models"
-
 
 def merge_diffusion_models(
     model_ids_or_paths: List[Union[str, Path]],
